components/training/requirements.txt
pandas==2.2.1
scikit-learn==1.4.1
google-cloud-storage==2.16.0
